TP_PARCIAL1/sistema_facturacion/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyswip==0.2.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
